pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dcheck-er"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
packages = ["dcheck_enterprise_runner"]
